refactor: move area chart to ui components
